Fix entities

+++ lib.php
@@ -793,7 +793,7 @@ function booking_update_options($optionvalues, $context) {
     // Every time we save an entity, we want to make sure that the name of the entity is stored in location.
     if (!empty($optionvalues->local_entities_entityid)) {
         // We might have more than one address, this will lead to more than one record which comes back.
-        if (class_exists('local/entitiesrelation_handler')) {
+        if (class_exists('local_entities\entitiesrelation_handler')) {
             $entities = entitiesrelation_handler::get_entity_by_id($optionvalues->local_entities_entityid);
             $option->address = '';
             foreach ($entities as $entity) {
@@ -1896,8 +1896,10 @@ function booking_delete_instance($id) {
 
     // Delete any entity relations for the booking instance.
     // TODO: this should be moved into delete_booking_option.
-    if (!entitiesrelation_handler::delete_entities_relations_by_bookingid($booking->id)) {
-        $result = false;
+    if (class_exists('local_entities\entitiesrelation_handler')) {
+        if (!entitiesrelation_handler::delete_entities_relations_by_bookingid($booking->id)) {
+            $result = false;
+        }
     }
 
     if (!$DB->delete_records("event", array("instance" => "$booking->id"))) {
